src/main/java/christmas/util/ConvertUtil.java
package christmas.util;

import christmas.domain.exception.ExceptionMessage;

public class ConvertUtil {
    private ConvertUtil() {
    }

    public static int convertToDay(final String input) {
        try {
            return Integer.parseInt(input);
        } catch (NumberFormatException e) {
            throw new IllegalArgumentException(ExceptionMessage.INVALID_DATE.getMessage());
        }
    }

    public static int convertToMenuNumber(final String input) {
        try {
            return Integer.parseInt(input);
        } catch (NumberFormatException e) {
            throw new IllegalArgumentException(ExceptionMessage.INVALID_ORDER.getMessage());
        }
    }
}
